Rename file to match compponent
